Separating platorm APIs
