+++ lib/clients/platform/index.js
@@ -221,6 +221,18 @@ class Platform {
       return Promise.reject(err);
     }
   }
+
+  async createPrivatehiveNetwork() {}
+
+  async listPrivatehiveNetworks() {}
+
+  async deletePrivatehiveNetwork() {}
+
+  /**
+   *
+   * @return {PrivatehiveNode}@memberof Privatehive
+   */
+  async getPrivatehiveNode() {}
 }
 
 module.exports = Platform;
